Librarify utransport_socket
Aligning the utransport_socket with recommendations from Pete. Instead of importing the utransport_socket.rs, we are now making it into a library crate.
