+++ .github/workflows/tck-test.yml
@@ -18,11 +18,6 @@ jobs:
     steps:
     - name: Checkout
       uses: actions/checkout@v3
-    - name: Initialize Rust library
-      run: |
-        cd up_client_socket/rust
-        cargo init --lib utransport-socket
-        cd ../../
     - name: Build Rust Test Agent
       uses: actions-rs/cargo@v1
       with:
