+++ test_agent/rust/Cargo.toml
@@ -22,21 +22,13 @@ exclude = [
     ".github/*",
     ".gitignore",
 ]
-# include = ["test_agent/rust/src/main.rs",
-#            "test_agent/rust/*"]
 
 [[bin]]
-name = "rust_tck"  # Specify the name of your executable
-# path = "test_agent/rust/src/main.rs"
+name = "rust_tck"
 
 [lints.clippy]
 all = "deny"
 pedantic = "deny"
-#cargo = "deny"
-#restriction = "deny"
-#nursery = "deny"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
 anyhow = "1.0.75"
@@ -49,11 +41,9 @@ prost = "0.12"
 prost-types = "0.12"
 protobuf = { version = "3.3" }
 utransport-socket = {path = "../../up_client_socket/rust/utransport-socket"}
-up-rust = { git = "https://github.com/eclipse-uprotocol/up-rust"} #rev = "68c8a1d94f0006daf4ba135c9cbbfddcd793108d" }
-#up-rust = { git = "https://github.com/PLeVasseur/uprotocol-rust_fork.git", rev = "9e3bc156c8b743f0f997d0b4d0888871e64c78ed"} 
+up-rust = { git = "https://github.com/eclipse-uprotocol/up-rust"}
 base64 = { version = "0.22.0"}
 serde = { version = "1.0", features = ["derive"] }
-#serde_json = { version = "1.0" }
 once_cell = "1.8.0" 
 tokio = { version = "1", features = ["full"] }
 anystruct = "0.1.0"
