+++ up_client_socket/rust/utransport-socket/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "utransport-socket"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0.75"
+async-std = "1.12.0"
+async-trait = "0.1"
+chrono = "0.4.31"
+env_logger = "0.10.0"
+log = "0.4.17"
+prost = "0.12"
+prost-types = "0.12"
+protobuf = { version = "3.3" }
+up-rust = { git = "https://github.com/eclipse-uprotocol/up-rust"}
+base64 = { version = "0.22.0"}
+serde = { version = "1.0", features = ["derive"] }
+once_cell = "1.8.0" 
+tokio = { version = "1", features = ["full"] }
+anystruct = "0.1.0"
+json2pb="*"
+protobuf-json-mapping = "3.4.0"
+serde_json = "1.0"
+uuid = "1.8.0"
